+++ src/views/errors/SpotifyError.vue
